Various tuning and bugfixes

+++ Workspace/GetARoom/src/main/java/com/soen343/JDBImappers/QueueNodeEdgeJDBIMapper.java
@@ -13,7 +13,7 @@ public class QueueNodeEdgeJDBIMapper implements ResultSetMapper<QueueNodeEdge> {
 
 	@Override
 	public QueueNodeEdge map(int index, ResultSet r, StatementContext ctx) throws SQLException {
-		 return new QueueNodeEdge(r.getLong("id"), r.getLong("parent_id"), r.getLong("child_id"));
+		 return new QueueNodeEdge(r.getInt("id"), r.getInt("parent_id"), r.getInt("child_id"));
 	}
 
 }
